==> analyis/00_setup.R <==
library(tidyverse)
library(lubridate)
library(openxlsx)
library(zoo)
library(data.table)
library(janitor)
library(epitools)

source("analyis/01_create_weekly.R")
source("analyis/02_summary_by_month.R")
source("analyis/03_loess_graphs.R")
source("analyis/04_draw_covid_cases.R")
source("analyis/05_summarise_cohort.R")
source("analyis/06_compare_pre_and_post_pandemic.R")
